requirements.txt
setuptools
pandas
numpy
pymysql
scikit-learn
python-dotenv
# -e .